requirement.txt
certifi==2019.11.28
email-split==1.0.0
MySQL-python==1.2.5
pymongo==3.10.1
PyYAML==5.3
sqlparse==0.1.16.dev0
